Split app component, init Auth component
