+++ client/src/components/Auth/Auth.js
@@ -1,0 +1,5 @@
+const Auth = () => {
+  const state = null;
+  return <div>AUTH</div>;
+};
+export default Auth;
